Remove redundant static_cast operations throughout codebase
This commit addresses clang-tidy readability-redundant-casting warnings
by removing unnecessary explicit casts where implicit conversion is safe
and appropriate:

- msg.cpp: Remove redundant uint8_t casts from bitwise OR operations with
  MIDI channel constants (7 instances)
- track.cpp: Remove redundant int casts from integer division operations
  (4 instances)
- fileread.cpp: Remove redundant int cast from return statement (1 instance)
- filewrite.cpp: Remove redundant int16_t casts from function parameters
  (3 instances)

All changes maintain type safety while improving code readability and
reducing verbosity. Build and all 114 unit tests pass successfully.

+++ src/jdksmidi_fileread.cpp
@@ -537,7 +537,7 @@ int MIDIFileRead::e_get_c()
     }
 
     --to_be_read;
-    return static_cast<int>(c);
+    return c;
 }
 
 void MIDIFileRead::msg_add(int a)
